Rename .reporting{.computations} → .report{.operator}
- Also rename test modules.
- Install DeprecatedPathFinder to support imports from old locations.
- Adjust imports minimally so package can be imported and tests pass.
