+++ message_ix/__init__.py
@@ -1,4 +1,5 @@
 import logging
+import sys
 from pathlib import Path
 
 try:
@@ -8,10 +9,11 @@ except ImportError:  # Python 3.7
 
 from ixmp import ModelError, config
 from ixmp.model import MODELS
+from ixmp.util import DeprecatedPathFinder
 
 from .core import Scenario
 from .models import MACRO, MESSAGE, MESSAGE_MACRO
-from .reporting import Reporter
+from .report import Reporter
 from .util import make_df
 
 __all__ = [
@@ -31,6 +33,18 @@ try:
 except PackageNotFoundError:  # pragma: no cover
     # Package is not installed
     __version__ = "999"
+
+# Install a finder that locates modules given their old/deprecated names
+sys.meta_path.append(
+    DeprecatedPathFinder(
+        __package__,
+        {
+            r"reporting(\..*)?": r"report\1",
+            "report.computations": "report.operator",
+        },
+    )
+)
+
 
 # Register configuration keys with ixmp core and set default
 config.register("message model dir", Path, Path(__file__).parent / "model")
