+++ message_ix/report/operator.py
