+++ message_ix/tests/report/test_operator.py
